api/requirements.txt
firebase_admin==6.0.1
Flask==2.2.2
Flask_Cors==3.0.10
protobuf==4.21.12
python-dotenv==0.21.0
python_bcrypt==0.3.2
python_magic==0.4.27
sendgrid==6.9.7
Werkzeug==2.2.2

google~=3.0.0
python-magic~=0.4.27
bcrypt~=4.0.1
opencv-python~=4.6.0.66
numpy~=1.23.5
cvlib~=0.2.7
Pillow~=9.3.0